Started working on collision in the sim, i have some good ideas
Basically we can check the intersection of each line of the rectangular
bounding box of the robot with the wall if the robot and the wall
actually intersect. Then we can get the point of intersection closest to
each wheel and use torque to figure out interactions. I'm not entirely
sure how it will play out, but hopefully it will work out in the end.